Refactor scripts tree structure
