+++ src/lib/kafka/producer.js
