+++ test/integration/settlementTransfer.test.js
@@ -41,7 +41,7 @@ const SettlementParticipantCurrencyModel = require('../../src/models/settlement/
 const TransferModel = require('@mojaloop/central-ledger/src/models/transfer/transfer')
 const TransferStateChangeModel = require('@mojaloop/central-ledger/src/models/transfer/transferStateChange')
 const ParticipantPositionModel = require('@mojaloop/central-ledger/src/models/position/participantPosition')
-const Producer = require('../../src/handlers/lib/kafka/producer')
+const Producer = require('../../src/lib/kafka/producer')
 // require('leaked-handles').set({ fullStack: true, timeout: 15000, debugSockets: true })
 
 const currency = 'USD'
